feat(coverage): extending coverage a bit
